requirements.txt
# Configure line endings for Windows
git config --global core.autocrlf true
pandas==2.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytest==7.4.2